Stop scheduling daily Sidekiq workers in local environment
I'd occasionally see a lot of Link Checker API activity locally,
and it seems to stem from this (in combination with me keeping
a `govuk-docker-up` shell running overnight).
We should avoid actioning this cronjob anywhere but on the remote
environments (integration, staging, production).

I did attempt to stop this happening in #10490, by adding an
`unless Rails.env.development?` clause around the config in the
`config/sidekiq.yml` file itself. But this was reverted in #10498
because it was causing issues with deployments, as the Rails
constant isn't defined when `config/sidekiq.yml` is evaluated.

This commit tweaks the approach to look for the `RAILS_ENV` variable,
which exists even if Rails itself is not initialised yet.