Omit lead images requiring cropping from news article content details
Now that image variants are only generated after cropping, we don't have an :s300 image for an uncropped lead image. This results in the URL to the image being nil.

Publishing API rejects lead image content with a null URL, so this was causing an error.

This commit omits lead images requiring cropping from the Publishing API content, and surfaces the issue in the lead image component.