Prevent null lead image error for case studies
Now that image asset variants are not generated until the image has been cropped, we need to make sure that we do not attempt to send lead image data to publishing API for images that are too large.

Sending image data for a missing variant results in it having a URL of nil, which is invalid in the Publishing API schema.