Only display groups where user can create at least one doc type
See https://github.com/alphagov/whitehall/pull/10804#discussion_r2499318723

We want to avoid displaying a group option if the current user
doesn't have permission to create any of its child document types.

In practice, we don't intend to surface this 'initial' screen to
users (when we switch on config-driven news articles for all,
we're going to redirect users straight to the news article
'interstitial' form) but it should be useful to have this logic
nonetheless.