Add 'body' method definition to StandardEdition [WHIT-2638]
There are various things in Whitehall that expect all editions to
have a `body` method, e.g.
https://github.com/alphagov/whitehall/blob/0ff06ba6736b686cd70fca107556b0a85214bb93/app/services/link_checker_api_service.rb#L14-L15

Without a `body` method definition, in the above example, we fail
to surface 'link check API reports' in the UI.

Whilst we could attempt to encode some sort of fall back to the
block_content["body"] property in all such cases, that could
quickly get messy as it would require us to also check the type
of document is a StandardEdition. It is safer to just define the
`body` method, delegating its value to the `block_content["body"]`
field if it exists (returning `nil` if it does not, which has much
the same effect as before this change).

The only consequence of this was having to specify a `body` value
in the standard edition translations controller test, otherwise
the NoFootnotesInGovspeakValidator would bork on `block_content: {}`
and not be able to call `=~` on a Hash.

Jira: https://gov-uk.atlassian.net/browse/WHIT-2638