Re-present migrated documents downstream to Publishing API
Whilst this is mostly an internal architectural rewrite with no
user visible changes, it is cleaner to make sure we update
Publishing API as there are some minor tweaks to the Publishing
API payload. Things like the unused `original_primary_publishing_organisation`
link will be dropped as a result.

No need to spin up a whole separate worker for the republishing
activity - let's do it as part of the same worker. Stops the
queues getting filled up with bulk_republish jobs.