Get sidekiq working locally [WHIT-2638]
I was confused as to why my workers in PR 10814 weren't being
processed - visiting `/sidekiq` locally (after enqueuing the
workers from the rails console) would show the workers in the
Enqueued queue, but never picked up.

It seems we didn't add the sidekiq startup invocation when
creating the Procfile.dev file in #8812. It doesn't look like it
was considered. I don't see a downside to enabling it, so have
done so here.

Jira: https://gov-uk.atlassian.net/browse/WHIT-2638