Route the new news article option through the standard editions controller
We're dispensing in this commit with the document view model, which is a pattern which never caught on in the rest of the codebase, and had the added disadvantage of not easily being able to leverage Rails path helpers.

The existing document type selection logic also had more behaviour than we strictly need at this point, so we simplify that down to just the group filtering.

At this point we have to remove the news article features. It isn't possible to create a news article via the UI any longer, so they have to go.