Add means of filtering out legacy & new content/links when diffing
We know the Publishing API presenter payloads will have SOME
differences when migrating from NewsArticle to StandardEdition.
For example, some legacy fields in the `content` payload will be
dropped - e.g. `details.first_public_at` - whereas we're obviously
also introducing _new_ fields in the `content` payload (such as
`details.image`, which is how we're setting the lead image). So
when diffing the payloads, we need a means of ignoring the known
legacy properties, and ignoring the known new properties.

The `ignore_legacy_content_fields` and `ignore_new_content_fields`
methods respectively accomplish that. There is also an equivalent
for the `links` payload: `ignore_legacy_links` and
`ignore_new_links`.