Add 'preview' page for changing document type
Step 2: following on from previous commit where we display the
available document types that can be switched to, this commit now
shows a preview of what that actually means in practice, breaking
down the 'properties' and 'associations' of each configurable
document type.

We've introduced a `DocumentTypeChangeSummary` component here, as
there are enough paths of code to test that it would have been
very noisy to try to do so through the controller tests.