Add 'Change' link to 'Type of document' row and write e2e test
Step 4: add the 'Change' link that is the entrypoint to the whole
journey of changing a configurable document's type.

Note that I've had to consolidate the "When I click" step definition
that already had a definition in the retagging steps, as otherwise
there was an ambiguous match.

Note also that we're currently missing test coverage for two minor
bits of view logic:

1. That the 'Change' link only appears on configurable documents...
2. ...that are in a draft state

It felt a bit overkill to write controller tests for the above,
and even more overkill to extract that bit of the UI into a view
component. I think the risk of that bit of code losing its
safeguards is pretty minimal, and in any case, if the Change link
were to appear on an Edition that it shouldn't appear on, it would
become pretty clear from subsequent screens that the user is
unable to proceed with a document type change.