Always send organisation crest to Publishing API
In 1f2abdc83630009c2ad033f8ec13e8ffb611f8e9 "no identity" and "custom" crest data was omitted from the Publishing API payload in order to fix an issue with Government Frontend. However, the crest data wasa not omitted from Search API, and the search API data is in fact needed by Collections frontend.

We are in the process of migrating Whitehall organisations on to the govuk index in Search API v1, which is indexed via Publishing API, so this presents a problem.

We are therefore reinstating the "no identity" and "custom" crest values in the Publishing payload and will fix any downstream issues this causes in the frontend.