Split out 'attachments' nav item and made conditionals consistent
The 'attachments' nav item was needlessly coupled to the
'standard edition nav items', breaking the convention we see
elsewhere in the file, where each category of thing is given its
own method.

Another convention that was not followed consistently was the use
of conditionals in the `edition_nav_items` method to decide whether
to call out to the other methods. This has now been applied in the
call to 'attachments_nav_items' and 'images_nav_items', simplifying
their implementations.