Avoid rendering <nav> element unless at least one tab exists
This component would continue to render a <nav> component with an
"Document navigation" aria-label even if there were no tabs to render, which could
be a problematic experience for screen reader users. So this bit
of logic conditionally renders the element only if there is at
least one tab to display.