Do not allow uncropped images in the select image block
Following the same pattern introduced in `f77d05b9`. The non-lead select block will allow svgs (which are not allowed as lead images), but should still not receive images that have not yet been cropped.

Since we filter what images we put in the block, there is no need to check them again in the pub api payload.