Make it possible to skip republishing and payload comparisons
Each worker was taking around 14 seconds to complete. We have
almost 104k documents to migrate, which would take 16.8 days to
run locally.

There may be times where we're confident enough in our recipes
that we can process the workers without doing the expensive payload
comparisons. We may also make the decision to prioritise a faster
Whitehall migration and defer the republishing of affected documents
to some other time.

1. Payload comparison was very useful during development, but once
   we'd run the migration across a few thousand news articles, it
   was unlikely to catch anything new. We've patched all of the
   little inconsistencies.
2. Representing to Publishing API is a nice aim, but we can also
   just settle for eventual consistency, or trigger a bulk republish
   of the document type as a whole once the migration has been
   completed. It doesn't necessarily have to be done one document
   at a time and slow down our migration.

We've therefore added two new keyword args to the
StandardEditionMigrator's `migrate!` method, to make all of the
documents in scope either require or not require republishing and
payload comparison. Note that these keyword args are converted to
simple JSON hash args when passed to the worker, to be compatible
with Sidekiq's API, which requires native JSON types, not symbols.