Allow focus on the first checkbox when there's an error on Excluded Nations
On the edition form, when there is a validation error on Excluded Nations, the error summary component links to the div that contains the checkboxes. However, the design system guidelines for [Error Summary component](https://design-system.service.gov.uk/components/error-summary/) says that when there's a group of checkboxes tied to a validation, we are to focus on the first checkbox. This also helps to prevent the default behaviour of updating the browser history with #[ELEMENT_ID] when the error item is clicked (keeping the history clean).

Consequently, I have introduced an explicit ID on the first checkbox, to make it easy for the error summary component to target it. In [error_summary_component.rb:41](https://github.com/alphagov/whitehall/blob/51e551308d6254e887e40f187ae6126f1ae92824/app/components/admin/error_summary_component.rb#L41), we can see that the error item derived its target ID  based on a combination of the parent and object classes. Hence, the explicit `edition_nation_inapplicabilities` in the first checkbox - This follows the example from the design system.

Also, to keep things consistent, I changed the ID of the section to `edition_nation_inapplicabilities_checkboxes`, because the explicit ID introduced to the first checkbox made the other checkboxes have random IDs (e.g `checkboxes-6f1be55d-1`). With that change, we have a fairly similar naming as before - `edition_nation_inapplicabilities_checkboxes-1`.

JIRA: https://gov-uk.atlassian.net/browse/WHIT-2747