Change `scope` to take array of documents rather than editions
The `editions` scope made sense earlier when we were specifying
which recipe those editions should follow, and when we had in
mind the idea of gradually migrating news articles by their
subtype rather than all in one go.

The code has evolved and now the recipes are determined on the
fly, and the migration has always been to a document overall
(inside a transaction) rather than editions in isolation. It now
makes much more sense to just pass an array of documents and avoid
the editions -> documents -> editions mapping that was happening
before.