Genericise the StandardEditionMigratorWorkerTest where possible
There were several references to concrete (legacy) news articles,
and in the spirit of enabling deleting code in future, we want
to remove all such references, referring to test types instead.

There was one test I wasn't able to find a straightforward
workaround for - the `migrate_editions!` test. I kept running into
issues with the Edition table single-table-inheritance, even when
defining test edition types such as `TestEdition < Edition`. For
now I've settled for linking it to the Publication content type,
which is a long way down the road on our migration journey (we'll
need to have a solution in place for multi page architecture first).