Filter duplicate role content IDs out of StandardEditionPresenter
There are a small handful of documents that have been associated
with more than one role appointment for the same role, which is
considered invalid:
https://github.com/alphagov/whitehall/blob/2e57781b61d2a4a36d28f0c601be7d0b59535688/app/models/role_appointment.rb#L48

It is not valid in Whitehall, nor is it valid in Publishing API:
attempting to send a payload containing duplicate role IDs causes
a Publishing API schema validation error.

In this commit I've carried over the behaviour already set in the
legacy news article presenter, which simply filters out any
duplicate role IDs at the point of sending to Publishing API:
https://github.com/alphagov/whitehall/blob/6a9dab86a24e1fdbed6d508dfbaff96469f19200/app/presenters/publishing_api/payload_builder/roles.rb#L25

This seems a harmless enough bit of defensive code that unblocks
us and helps us avoid tackling the thornier issue of whether or
not duplicate role IDs _should_ be allowed (on 'overlapping' days).
The result was inconclusive when asking Content colleagues:
https://gds.slack.com/archives/CACV3TACU/p1764002238414669

For posterity, for the offending documents (scoped to just News
Articles) are as follows. (There may be others that are not news
articles).

```
Document ID, Edition ID, Content ID, slug
4525, 191418, "5c769c11-7631-11e4-a3cb-005056011aef", "putting-aviation-red-tape-under-the-spotlight",
253461, 371774, "5fe40119-7631-11e4-a3cb-005056011aef", "ministerial-appointments-august-2014",
318759, 584658, "19d655a8-b8fa-4074-97d5-9a96e050b4fb", "uk-and-poland-discuss-closer-co-operation",
324352, 605161, "1b74df55-3f9f-4db1-bc58-ede097130322", "new-ministerial-appointments-19-march-2016",
370479, 781674, "1d4f79a0-4b97-48a2-87b7-c513d9043792", "ministerial-appointments-2-november-2017",
473182, 1267351, "3ec1d32a-75ba-4b36-b971-69ec2d966f6c", "uk-to-host-its-first-global-lgbt-conference",
```

Jira: https://gov-uk.atlassian.net/browse/WHIT-2487