Add date validator
We are not able to use the `date_attributes` method because it is a class method and, in this case of config-driven types, we would not know what the attributes are at class definition time. Therefore, we have to use the `pre_validate_date_attribute` function directly in the block content attribute setter.