Add `url_to_document_slug` helper method
Sometimes we want to grab the bit of the URL that corresponds to
what the 'slug' would be on the Document. It's usually just the
last bit of the base path, but in the case of HTML attachments,
it's the bit of the base path before that, and then Statistics
Announcements are their own special case. Hence the need for a
helper!

This will be used in the next commit.