WIP - trying to figure out how to do conditional formatting
We ended the day discussing alternatives, e.g. ripping out validation
from the config altogether and just defining a validator for each
config driven document type. The validation is in many ways
separate from the definition of the properties etc and it would
pave the way for testing the validation in isolation like we've
attempted to start to do with the topical events validations.