WIP: moving 'complex' validation out to its own custom validator
See previous commit. Trying to represent fragmented and conditional
validation rules that are dependent on two or more distinct
properties, in pure JSON, and then mapping that to our generic
block based ruby framework is... nightmarish.

And for not much gain.

Much simpler to just declare a content-type-specific validator and
reference that from the JSON. We can encode all manner of validation
rules in that.

TODO: would we want to move the 'simpler' validation rules out
to a content-type-specific validator too? For consistency, yes,
but also seems a shame when the existing content types have been
able to get away without it so far.