Fix existing cucumber test for adding a Welsh translation
Previous test read as though it was stepping through the process
of creating a translation - that the input fields were pre populated
with primary locale content, and so on - whereas in reality the
translation was already created and saved at the point of
checking the primary locale defaults. The test also needlessly
had a method defined when it was the only occurrence of that
method - I've now inlined the code.