Fix bug where translated block content was being lost on new draft
We had only written tests for - and manually tested - the scenario
of going to create a translation on a draft document and checking
that our inputted translations persisted on save. We had no tests
for ensuring that translated block content is carried over through
the edition lifecycle. As it happens, the Edition::Translatable
`process_associations_before_save` callback was only carrying
over title, summary and body - not block_content - so translated
block content was being automatically lost whenever a new draft
of a document was created.

This commit fixes that bug, adding a unit test to cover the
behaviour. I had also written a feature test in the act of fixing
the bug, as I wasn't sure initially where in the codebase the bug
lay. We could now delete the feature test if we want to, but for
such an important feature I think it's best to go with the belt
and braces approach.