Add self link to organisation Publishing API payload
We are in the process of having organisations be indexed in Search API via Publishing API. There is a search use case for allowing the organisation page to appear in the results when filtering for all documents belonging to the organisation. This relies on the page having a link to the organisation, so we need to make sure this link is present in the payload, even if it is circular.