Add new block content abstraction to present config-driven blocks
This now replaces the need for all the `publishing_api_payload` methods defined on all the config driven blocks. There is no need to define a payload behaviour for the `default_object` block, as we always send a flat details hash to publishing API. Only "leaf" blocks will be included in the payload.

We plan to maintain both implementations in the code to de-risk the release. At a later point, we will be removing all the `publishing_api_payload` methods and only use the new flow.

All the tests covering the blocks' `publishing_api_payload` behaviour have been ported over.