Introduce alternative way to compute standard edition presenter payload
We want to specify which attributes we're going to be including in the presenter, separately from defining the model attributes, or the forms to render.

Note that the check for empty on the details hash is not very precise. We could in theory have blank `details`, if there is no content to send, or if the `presenters` hash is blank (does not specify any `attributes`), but in reality we always require at least one block in the details hash for our existing types.