Do not attempt to clear the redirect on deleted assets
After an edition is published, Whitehall clears any redirects for the attachment assets from Asset Manager. It does this so that if the document was previously unpublished, users will be able to access the assets again.

However, in the case where the asset has been deleted from the newly published edition, this is unnecessary, because the asset should no longer be accessible.

When an edition is published, an asset deletion worker runs for each deleted attachment marking the asset as deleted in Asset Manager. If this completes before the Whitehall redirect worker, then the request to clear the redirect URL will be rejected. Currently in this situation we raise an exception in the Asset Updater. This has resulted in lots of noisy errors.

This commit avoids the error by preventing the attempt to clear redirects for aattachments that have been deleted.