Do not raise the asset deleted error if updating a draft deleted asset
This is actually a valid use case. When the draft edition associated with a deleted attachment that owns this asset is published, we want to update the draft state of the asset to false and update the parent document URL to a non-draft origin URL.

If we don't update the asset to live, and the deleted asset was a replacement for another asset, then the replaced asset would remain live - a serious bug. Asset Manager only applies the replacement if the replacing asset is live.

The code runs separate operations to achieve this update+delete scenario:

- when the edition (with a deleted attachment) is published the edition services call the `AttachmentAssetPublisher`, which enqueues `PublishAttachmentAssetJob`-> which in turn DELETES the replacement asset (`AssetDeleter`). The original asset gets set with a replacement ID of the replacement asset.
- a round of updates then sort out the draft state of the replacement via the `AssetManager::AssetUpdater`. At this point in time the asset IS deleted in AM, and we only want to update its draft state to false, so the redirect between the assets works as expected.

A breakdown of the flow:
- edition with file gets published
- new draft is made
- the file is replaced
- the replacement is deleted
- the edition gets published again

Desired state in Asset Manager after these actions is:
- original asset - Draft: false, Deleted: false, Replaced: true (live and not deleted, but redirecting)
- replacement asset - Draft: false, Deleted: true

The effect is that both assets are GONE.