Make use of 'display_type' consistent across editionable types
- Why: the display type key will be used to render the "type" of documents in tables/feature tables/dashboard in upcoming commits, as well as making the render consistent on the edition summary page (see fixed edition helper tests).
- Refactor usage of `display_type` to be computed following the same pattern in publications, speeches and news articles. If a type is not set, we fall back to the "genus".
- Remove opinionated computations of the display type so that all legacy models fetch it from locale values, which is a clear source of truth. Standard edition overrides the method directly and returns the schema title.

Speech, News Article, Publication, and Corporate information pages, all override the `display_key` used by the `display_type`. The `display_key` is then used to fetch the correct human-readable subtype label from the locale file.

Speech changes:
- Written and oral statements to parliament were both displayed as "Statement to Parliament" -> now changed to display "Written statement to Parliament" and "Oral statement to Parliament", respectively.
- Removed the generic "statement_to_parliament" key from locale as no longer in use.

StandardEdition makes use of the label (schema title) set on the type config, so it does not need to use the locale. It overrides the `display_type` method directly.

Case study does not need to override the `display_type_key`, it can compute it from `format_name`.

Call for evidence and consultation have their own logic for the key, based on various states (though not really a type/subtype scenario, it behaves the same in terms of rendering).

StatisticalDataSet and WorldwideOrganisation do not need to override the display key.

Detailed guide
- Removed the format_name "detailed guidance" -> we can use the generic format_name "detailed_guide". This does not seem to be used anywhere downstream. It is used in some mail notifications but only as subject.
- They generated `display_type_key` is now "detailed_guide" and has been update in the locale key to match.

Content changes:
- DocumentCollection's locale value changed to "Document collection".
- DetailedGuide locale value to "Detailed guide". This now implicitly changes how the documents index page filter option for this type displays. Changed from "Detailed guidance" -> "Detailed guide" (which matches how we refer to this type elsewhere).