Remove coverage for news article from feature tests
Since news articles will be removed from the codebase, we should update the tests to use a different type.

If we need to expand the coverage for standard edition, we can do it separately. For now, it suffices to keep the coverage for the as is, using an alternative document type, e.g. publication.

Co-authored-by: Alex Newton <alex.newton@digital.cabinet-office.gov.uk>