Remove coverage for news article from features
Since news articles will be removed from the codebase, we should update the tests to use a different type.

If we need to expand the coverage for standard edition, we can do it separately. For now, it suffices to keep the coverage for the as is, using an alternative document type, e.g. publication.

Differences from previous PR:
- no need to complicate it with type checks - we just don't want to run the same feature tests for old and new worlds - better to duplicate and be explicit what features we will cover for standard edition -> and add them to standard-edition.feature or new test files. That would be out of scope for this pr though.