[WHIT-2724] Fix minor inconsistencies in block_content
Make the data that is stored in block_content consistent with the original translations that it
is being copied from. Instead of directly getting the values from the edition, we instead use
the translation, so that we avoid getting fallback values and including data that is not part
of the original translation, or only relevant to the primary translation.

Also ensures we can't save anything in the `body` column for a StandardEdition, as we only care
about the value in block_content. This will hopefully prevent any confusion about where values
originate from.